+++ kamehouse-shell/bin/kamehouse/deploy-kamehouse.sh
@@ -24,8 +24,8 @@ if [ "$?" != "0" ]; then
   echo -e "\033[1;36m$(date +%Y-%m-%d' '%H:%M:%S)\033[0;39m - [\033[1;31mERROR\033[0;39m] - \033[1;31mAn error occurred importing deployment-functions.sh\033[0;39m"
   exit 99
 fi
-loadKamehouseSecrets
 
+LOAD_KAMEHOUSE_SECRETS=true
 EXIT_CODE=${EXIT_SUCCESS}
 DEPLOYMENT_DIR=""
 FAST_BUILD=true