+++ kamehouse-shell/bin/kamehouse/docker/docker-ci-integration-tests-trigger.sh
@@ -11,8 +11,8 @@ if [ "$?" != "0" ]; then
   echo -e "\033[1;36m$(date +%Y-%m-%d' '%H:%M:%S)\033[0;39m - [\033[1;31mERROR\033[0;39m] - \033[1;31mAn error occurred importing docker-functions.sh\033[0;39m"
   exit 99
 fi
-loadKamehouseSecrets
 
+LOAD_KAMEHOUSE_SECRETS=true
 INTEGRATION_TESTS_SUCCESS_MESSAGE="SUCCESS EXECUTING INTEGRATION TESTS"
 SCRIPT="kamehouse/docker/docker-container/docker-integration-tests-run.sh"
 DOCKER_PORT_HTTP=${DOCKER_PORT_HTTP_CI}