+++ kamehouse-shell/bin/kamehouse/docker/docker-run-kamehouse.sh
@@ -12,8 +12,8 @@ if [ "$?" != "0" ]; then
   echo -e "\033[1;36m$(date +%Y-%m-%d' '%H:%M:%S)\033[0;39m - [\033[1;31mERROR\033[0;39m] - \033[1;31mAn error occurred importing docker-functions.sh\033[0;39m"
   exit 99
 fi
-loadKamehouseSecrets
 
+LOAD_KAMEHOUSE_SECRETS=true
 BUILD_ON_STARTUP=false
 BUILD_ON_STARTUP_PARAM=""
 DEBUG_MODE=false