+++ kamehouse-shell/bin/kamehouse/kamehouse-mobile-resync-static-files.sh
@@ -6,8 +6,8 @@ if [ "$?" != "0" ]; then
   echo -e "\033[1;36m$(date +%Y-%m-%d' '%H:%M:%S)\033[0;39m - [\033[1;31mERROR\033[0;39m] - \033[1;31mAn error occurred importing kamehouse-functions.sh\033[0;39m"
   exit 99
 fi
-loadKamehouseSecrets
 
+LOAD_KAMEHOUSE_SECRETS=true
 DELETE_ONLY=false
 SOURCE_FILES_KAMEHOUSE_DIR=""
 SOURCE_FILES_GROOT_DIR=""