+++ kamehouse-shell/bin/kamehouse/mariadb-archive-tennisworld-bookings.sh
@@ -6,9 +6,8 @@ if [ "$?" != "0" ]; then
   echo -e "\033[1;36m$(date +%Y-%m-%d' '%H:%M:%S)\033[0;39m - [\033[1;31mERROR\033[0;39m] - \033[1;31mAn error occurred importing kamehouse-functions.sh\033[0;39m"
   exit 99
 fi
-loadKamehouseSecrets
 
-# Global variables
+LOAD_KAMEHOUSE_SECRETS=true
 MARIADB_ARCHIVE_FILE=${HOME}/programs/kamehouse-shell/sql/mariadb/archive-tennisworld-bookings.sql
 
 mainProcess() {