+++ kamehouse-shell/bin/kamehouse/stop-kamehouse.sh
@@ -6,9 +6,8 @@ if [ "$?" != "0" ]; then
   echo -e "\033[1;36m$(date +%Y-%m-%d' '%H:%M:%S)\033[0;39m - [\033[1;31mERROR\033[0;39m] - \033[1;31mAn error occurred importing kamehouse-functions.sh\033[0;39m"
   exit 99
 fi
-loadKamehouseSecrets
 
-# Initial config
+LOAD_KAMEHOUSE_SECRETS=true
 OPERATION=stop
 
 mainProcess() {