+++ changelog.md
@@ -7,6 +7,8 @@
 
 #### v10.03
 
+- Removed fallback to unencrypted kamehouse secrets file
+- Added pre parse cmd arguments function to shell scripts framework
 - Refactored functions to execute ssh, scp and sftp commands from kamehouse shell
 - Refactored mobile backend servers configuration to use more explicit values
 - Refactored kamehouse servers list configuration to use more explicit values